Split script into modules
